+++ src/net-util/tests/HttpHeaders.test.js
@@ -20,17 +20,20 @@ describe('constructor', () => {
   });
 });
 
+// This tests the constructor and methods in terms of modifying an initially
+// empty instance.
 describe.each`
-label            | constructor
-${'constructor'} | ${true}
-${'appendAll()'} | ${false}
-`('$label', ({ constructor }) => {
+label            | methodName
+${'constructor'} | ${'constructor'}
+${'appendAll()'} | ${'appendAll'}
+${'setAll()'}    | ${'setAll'}
+`('$label', ({ methodName }) => {
   function prep(arg) {
-    if (constructor) {
+    if (methodName === 'constructor') {
       return new HttpHeaders(arg);
     } else {
       const hh = new HttpHeaders();
-      hh.appendAll(arg);
+      hh[methodName](arg);
       return hh;
     }
   }
@@ -76,6 +79,16 @@ ${'appendAll()'} | ${false}
       expect([...hh]).toEqual([['set-cookie', 'x=1'], ['set-cookie', 'y=2']]);
       expect(hh.getSetCookie()).toIncludeSameMembers(['x=1', 'y=2']);
     });
+  });
+
+  test('includes all values in an array', () => {
+    const hh = prep({ 'foo': ['bar', 'baz'] });
+    expect([...hh]).toEqual([['foo', 'bar, baz']]);
+  });
+
+  test('includes all values from multiple entries with the same name', () => {
+    const hh = prep([['foo', 'bar'], ['foo', 'baz']]);
+    expect([...hh]).toEqual([['foo', 'bar, baz']]);
   });
 });
 
@@ -150,7 +163,7 @@ describe('appendAll()', () => {
     ]);
   });
 
-  test('uses an underlay function when it would not overwrite an existing header', () => {
+  test('uses an underlay function when it would not add to a pre-existing header', () => {
     const hh = new HttpHeaders();
 
     hh.appendAll({ 'blorp': () => 'beep-boop' });
@@ -158,7 +171,7 @@ describe('appendAll()', () => {
     expect([...hh]).toEqual([['blorp', 'beep-boop']]);
   });
 
-  test('does not use the underlay function when it would overwrite an existing header', () => {
+  test('does not use the underlay function when it would add to a pre-existing header', () => {
     const hh            = new HttpHeaders();
     let   overlayCalled = false;
 
@@ -173,6 +186,21 @@ describe('appendAll()', () => {
 
     expect([...hh]).toEqual([['foo', 'bar']]);
     expect(overlayCalled).toBeFalse();
+  });
+
+  test('uses multiple underlay functions when they would not add to a pre-existing header', () => {
+    const hh = new HttpHeaders();
+
+    // This tests multiple entries with the same name.
+    hh.appendAll([['blorp', () => 'beep1'], ['blorp', () => 'beep2']]);
+
+    // This tests a single array entry binding to an array of functions.
+    hh.appendAll({ zorch: [() => 'zonk1', () => 'zonk2'] });
+
+    expect([...hh]).toIncludeSameMembers([
+      ['blorp', 'beep1, beep2'],
+      ['zorch', 'zonk1, zonk2']
+    ]);
   });
 });
 
@@ -323,5 +351,115 @@ ${'hasAny'}
     });
 
     expect(hh[methodName](...args)).toBe(expected);
+  });
+});
+
+describe('setAll()', () => {
+  test('sets a non-existent name to a single given value', () => {
+    const hh = new HttpHeaders();
+
+    hh.set('a', 'b');
+    hh.setAll({ 'x': 'y' });
+
+    expect([...hh]).toIncludeSameMembers([['a', 'b'], ['x', 'y']]);
+  });
+
+  test('sets a non-existent name to multiple given values', () => {
+    const hh = new HttpHeaders();
+
+    hh.set('a', 'b');
+    hh.setAll({ 'x': ['y', 'z'] });
+
+    expect([...hh]).toIncludeSameMembers([['a', 'b'], ['x', 'y, z']]);
+  });
+
+  test('replaces a pre-existing name with a single value', () => {
+    const hh = new HttpHeaders();
+
+    hh.set('a', 'b');
+    hh.set('x', 'y');
+    hh.setAll({ 'x': 'z' });
+
+    expect([...hh]).toIncludeSameMembers([['a', 'b'], ['x', 'z']]);
+  });
+
+  test('replaces a pre-existing name with multiple values', () => {
+    const hh = new HttpHeaders();
+
+    hh.set('a', 'b');
+    hh.set('x', 'y');
+    hh.append('x', 'z');
+    hh.setAll({ 'x': ['z1', 'z2'] });
+
+    expect([...hh]).toIncludeSameMembers([
+      ['a', 'b'], ['x', 'z1, z2']
+    ]);
+  });
+
+  test('sets multiple `set-cookies` from a plain object', () => {
+    const hh = new HttpHeaders();
+
+    hh.set('set-cookie', 'a=1');
+    hh.setAll({ 'set-cookie': ['b=2', 'c=3'] });
+
+    expect([...hh]).toIncludeSameMembers([
+      ['set-cookie', 'b=2'],
+      ['set-cookie', 'c=3']
+    ]);
+  });
+
+  test('sets multiple `set-cookies` from a `Headers`', () => {
+    const hh   = new HttpHeaders();
+    const orig = new Headers();
+
+    hh.set('set-cookie', 'a=1');
+    orig.append('set-cookie', 'b=2');
+    orig.append('set-cookie', 'c=3');
+    hh.setAll(orig);
+
+    expect([...hh]).toIncludeSameMembers([
+      ['set-cookie', 'b=2'],
+      ['set-cookie', 'c=3']
+    ]);
+  });
+
+  test('uses an underlay function when it would not replace to a pre-existing header', () => {
+    const hh = new HttpHeaders();
+
+    hh.setAll({ 'blorp': () => 'beep-boop' });
+
+    expect([...hh]).toEqual([['blorp', 'beep-boop']]);
+  });
+
+  test('does not use the underlay function when it would replace a pre-existing header', () => {
+    const hh            = new HttpHeaders();
+    let   overlayCalled = false;
+
+    hh.set('foo', 'bar');
+
+    hh.setAll({
+      'foo': () => {
+        overlayCalled = true;
+        return 'beeeeeeeep!!!!!';
+      }
+    });
+
+    expect([...hh]).toEqual([['foo', 'bar']]);
+    expect(overlayCalled).toBeFalse();
+  });
+
+  test('uses multiple underlay functions when they would not replace a pre-existing header', () => {
+    const hh = new HttpHeaders();
+
+    // This tests multiple entries with the same name.
+    hh.setAll([['blorp', () => 'beep1'], ['blorp', () => 'beep2']]);
+
+    // This tests a single array entry binding to an array of functions.
+    hh.setAll({ zorch: [() => 'zonk1', () => 'zonk2'] });
+
+    expect([...hh]).toIncludeSameMembers([
+      ['blorp', 'beep1, beep2'],
+      ['zorch', 'zonk1, zonk2']
+    ]);
   });
 });